Add Prime Vue + theming

+++ app.go
@@ -5,12 +5,10 @@ import (
 	"fmt"
 )
 
-// App struct
 type App struct {
 	ctx context.Context
 }
 
-// NewApp creates a new App application struct
 func NewApp() *App {
 	return &App{}
 }
@@ -23,7 +21,7 @@ func (a *App) startup(ctx context.Context) {
 
 type Person struct {
 	Name    string   `json:"name"`
-	Age     uint8    `json:"age"`
+	Age     uint32   `json:"age"`
 	Address *Address `json:"address"`
 }
